+++ package.json
@@ -3,7 +3,7 @@
 	"publishConfig": {
 		"registry": "https://npm.pkg.github.com/hiimlex"
 	},
-	"version": "0.1.0",
+	"version": "0.1.4",
 	"description": "React components library",
 	"scripts": {
 		"rollup": "rollup -c",
