+++ src/components/LeButton/index.ts
@@ -1,1 +1,2 @@
+export { default } from "./LeButton";
 export * from "./LeButton";
