+++ src/types/components/LeButton.ts
@@ -1,0 +1,13 @@
+export type LeButtonVariants =
+	| "primary"
+	| "secondary"
+	| "success"
+	| "danger"
+	| "warning"
+	| "default";
+
+export type LeButtonTypes = "filled";
+
+export interface LeButtonState {
+	disabled?: boolean;
+}
