+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./generics";
+export * from "./components";
